requirements.txt
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
blobfile==2.1.1
causality-llm==1.0
causallib==0.9.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
data==0.4
dataclasses-json==0.6.3
datasets==2.16.1
decorator==5.1.1
dill==0.3.7
diskcache==5.6.3
distro==1.8.0
docker-pycreds==0.4.0
exceptiongroup==1.2.0
fairscale==0.4.13
filelock==3.13.1
fire==0.5.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
funcsigs==1.0.2
future==1.0.0
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.3
hydra-core==1.3.2
idna==3.6
importlib-resources==6.1.1
interegular==0.3.3
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lark==1.1.9
latex==0.7.0
llama==0.0.1
llama3==0.0.1
llvmlite==0.42.0
lxml==4.9.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.0
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.3.8
packaging==23.2
pandas==2.1.3
patsy==0.5.4
Pillow==10.1.0
pip==23.3.1
protobuf==4.25.2
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycryptodomex==3.20.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.17.1
safetensors==0.4.2
scikit-learn==1.1.3
scikit-uplift==0.5.1
scipy==1.11.4
sentence-transformers==2.3.1
sentencepiece==0.1.99
sentry-sdk==1.40.4
setproctitle==1.3.3
setuptools==68.0.0
shutilwhich==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tempdir==0.7.1
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.19.1
torch==2.1.1
tqdm==4.66.1
transformers==4.40.1
triton==2.1.0
tueplots==0.0.15
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.3
wget==3.2
wheel==0.41.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
